requirements.txt
google_api_python_client==2.133.0
httplib2==0.22.0
moviepy==1.0.3
numpy==1.26.4
oauth2client==4.1.3
pedalboard==0.9.6
python-dotenv==1.0.1
redis==5.0.5
Requests==2.32.3
soundfile==0.12.1
yt_dlp==2024.07.01
